ESP -> Server connection

+++ Client/main.py
@@ -22,17 +22,22 @@ import gc
 def measure_temp():
     try:
         ds.convert_temp()
-        return round(ds.read_temp(roms[0]), 2)
+        return round(ds.read_temp(roms[0]), cfg["temp_precision_places"])
     except:
-        return None
+        return -124101.25  #none
     
 
 
 ########################################################
 def broadcastData(data):
-    print(data)     # 'Temperature: {}°C'.format(temp)
     client.publish(cfg["mqtt"]["temp_topic"], json.dumps(data), retain=False, qos=0)
 
+###########################################################
+
+def getISOTime():
+    t = time.gmtime()
+    return "%d-%02d-%02dT%02d:%02d:%09.6f" % (t[0], t[1], t[2], t[3], t[4], t[5])
+   
 ###########################################################
 
 
@@ -53,6 +58,7 @@ client = MQTTClient(CLIENT_ID, cfg["mqtt"]["broker"], cfg["mqtt"]["port"], cfg["
 
 #client.set_last_will(topic, msg, retain=False, qos=0)
 #client.set_callback(on_receive)
+# 'Temperature: {}°C'.format(temp)
 
 
 ######################################## Network
@@ -71,7 +77,7 @@ time.sleep(10)
 try:
     if wlan.isconnected(): 
         print("Connected to WIFI")
-        #ntptime.settime()
+        ntptime.settime()
         client.connect(clean_session=True)
 		
     else: print("Not connected to WIFI")
@@ -87,42 +93,47 @@ roms = ds.scan()
 
 #########################################
 while True:
-    dataStr = {"team_name": "orange", "created_on":time.time(), "temperature": measure_temp()}
+    startTime = time.ticks_ms()
+    m_temp = measure_temp()
+    dataStr = {"team_name": cfg["team"], "created_on":getISOTime(), "temperature": m_temp}
+    print(dataStr)   
+    
     
     if last_WLAN_state != wlan.status():
         print("WLAN status changed from {} to {}".format(last_WLAN_state, wlan.status()))
         last_WLAN_state = wlan.status()
 
-    
-    try:
-            if has_saved_data:
-                dataFile = open("hist.txt", 'r')
-                data = dataFile.readlines()
-                dataFile.close()
+    if m_temp is not None:
+        try:
+                if has_saved_data:
+                    print("Sending saved data...")
+                    dataFile = open("hist.txt", 'r')
+                    data = dataFile.readlines()
+                    dataFile.close()
 
-                for line in data:
-                    broadcastData(line)
-                
-                broadcastData(dataStr)
-                
-                os.remove("hist.txt")
-                has_saved_data = False
-                gc.collect()
-            else:
-                broadcastData(dataStr)
-    except: 
-        print("Can't send data, saving for later...") 
-        temp_history = open("hist.txt", "a")
-        temp_history.write(json.dumps(dataStr))
-        temp_history.write('\n')
-        temp_history.close()
+                    for line in data:
+                        broadcastData(line)
+                    
+                    broadcastData(dataStr)
+                    
+                    os.remove("hist.txt")
+                    has_saved_data = False
+                    gc.collect()
+                else:
+                    broadcastData(dataStr)
+        except: 
+            print("Can't send data, saving for later...") 
+            temp_history = open("hist.txt", "a")
+            temp_history.write(json.dumps(dataStr))
+            temp_history.write('\n')
+            temp_history.close()
 
-        has_saved_data = True
+            has_saved_data = True
 
-		###
-        fileToRead = open("hist.txt", 'r')
-        print("Reading: ", fileToRead.readlines())
-        fileToRead.close()
+            ###
+            #fileToRead = open("hist.txt", 'r')
+            #print("Reading: ", fileToRead.readlines())
+            #fileToRead.close()
 
-
-    time.sleep(cfg["update_period"])
+    print(time.ticks_ms() - startTime)
+    time.sleep_ms(cfg["update_period"] * 1000 - time.ticks_ms() + startTime)
